+++ R/parser.R
@@ -177,12 +177,12 @@ valid_formats <- c("xPONENT", "INTELLIFLEX")
 #' @return Plate file containing the Luminex data
 #'
 #' @examples
-#' plate_file <- system.file("extdata", "CovidOISExPONTENT.csv", package = "SerolyzerR")
-#' layout_file <- system.file("extdata", "CovidOISExPONTENT_layout.csv", package = "SerolyzerR")
+#' plate_file <- system.file("extdata", "CovidOISExPONTENT.csv", package = "SerolyzeR")
+#' layout_file <- system.file("extdata", "CovidOISExPONTENT_layout.csv", package = "SerolyzeR")
 #' plate <- read_luminex_data(plate_file, layout_file)
 #'
-#' plate_file <- system.file("extdata", "CovidOISExPONTENT_CO.csv", package = "SerolyzerR")
-#' layout_file <- system.file("extdata", "CovidOISExPONTENT_CO_layout.xlsx", package = "SerolyzerR")
+#' plate_file <- system.file("extdata", "CovidOISExPONTENT_CO.csv", package = "SerolyzeR")
+#' layout_file <- system.file("extdata", "CovidOISExPONTENT_CO_layout.xlsx", package = "SerolyzeR")
 #' # To suppress warnings and additional information use verbose = FALSE
 #' plate <- read_luminex_data(plate_file, layout_file, verbose = FALSE)
 #'