+++ tests/testthat/test-get-nmfi.R
@@ -2,8 +2,8 @@ library(testthat)
 
 test_that("get_nmfi works on our data with multiple parameters", {
   # Read plate
-  path <- system.file("extdata", "CovidOISExPONTENT.csv", package = "SerolyzerR", mustWork = TRUE)
-  layout_path <- system.file("extdata", "CovidOISExPONTENT_layout.xlsx", package = "SerolyzerR", mustWork = TRUE)
+  path <- system.file("extdata", "CovidOISExPONTENT.csv", package = "SerolyzeR", mustWork = TRUE)
+  layout_path <- system.file("extdata", "CovidOISExPONTENT_layout.xlsx", package = "SerolyzeR", mustWork = TRUE)
 
   expect_no_error(plate <- read_luminex_data(path, format = "xPONENT", layout_filepath = layout_path, verbose = FALSE))
 
@@ -14,8 +14,8 @@ test_that("get_nmfi works on our data with multiple parameters", {
 })
 
 test_that("get_nmfi with incorrect params", {
-  path <- system.file("extdata", "CovidOISExPONTENT.csv", package = "SerolyzerR", mustWork = TRUE)
-  layout_path <- system.file("extdata", "CovidOISExPONTENT_layout.xlsx", package = "SerolyzerR", mustWork = TRUE)
+  path <- system.file("extdata", "CovidOISExPONTENT.csv", package = "SerolyzeR", mustWork = TRUE)
+  layout_path <- system.file("extdata", "CovidOISExPONTENT_layout.xlsx", package = "SerolyzeR", mustWork = TRUE)
 
   expect_no_error(plate <- read_luminex_data(path, format = "xPONENT", layout_filepath = layout_path, verbose = FALSE))
 
@@ -28,8 +28,8 @@ test_that("get_nmfi with incorrect params", {
 
 
 test_that("get_nmfi on artificial plate", {
-  path <- system.file("extdata", "random.csv", package = "SerolyzerR", mustWork = TRUE)
-  layout_path <- system.file("extdata", "random_layout.xlsx", package = "SerolyzerR", mustWork = TRUE)
+  path <- system.file("extdata", "random.csv", package = "SerolyzeR", mustWork = TRUE)
+  layout_path <- system.file("extdata", "random_layout.xlsx", package = "SerolyzeR", mustWork = TRUE)
 
   expect_no_error(plate <- read_luminex_data(path, format = "xPONENT", verbose = FALSE))