+++ tests/testthat/test-process-dir.R
@@ -1,9 +1,9 @@
 library(testthat)
 
 test_that("Test finding layout file", {
-  plate_filepath <- system.file("extdata", "CovidOISExPONTENT_CO.csv", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
-  random_plate_filepath <- system.file("extdata", "random_intelliflex.csv", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
-  layout_filepath <- system.file("extdata", "CovidOISExPONTENT_CO_layout.xlsx", package = "SerolyzerR", mustWork = TRUE)
+  plate_filepath <- system.file("extdata", "CovidOISExPONTENT_CO.csv", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
+  random_plate_filepath <- system.file("extdata", "random_intelliflex.csv", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
+  layout_filepath <- system.file("extdata", "CovidOISExPONTENT_CO_layout.xlsx", package = "SerolyzeR", mustWork = TRUE)
 
   output <- find_layout_file(plate_filepath)
   expect_true(check_path_equal(output, layout_filepath))
@@ -11,11 +11,11 @@ test_that("Test finding layout file", {
 })
 
 test_that("Test checking for mba file", {
-  plate1_filepath <- system.file("extdata", "CovidOISExPONTENT.csv", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
-  plate2_filepath <- system.file("extdata", "CovidOISExPONTENT_CO.csv", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
-  plate3_filepath <- system.file("extdata", "random_intelliflex.csv", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
-  non_matched_filepath <- system.file("extdata", "random.csv", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
-  layout_filepath <- system.file("extdata", "CovidOISExPONTENT_CO_layout.xlsx", package = "SerolyzerR", mustWork = TRUE)
+  plate1_filepath <- system.file("extdata", "CovidOISExPONTENT.csv", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
+  plate2_filepath <- system.file("extdata", "CovidOISExPONTENT_CO.csv", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
+  plate3_filepath <- system.file("extdata", "random_intelliflex.csv", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
+  non_matched_filepath <- system.file("extdata", "random.csv", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
+  layout_filepath <- system.file("extdata", "CovidOISExPONTENT_CO_layout.xlsx", package = "SerolyzeR", mustWork = TRUE)
 
   expect_true(is_mba_data_file(plate1_filepath))
   expect_true(is_mba_data_file(plate2_filepath))
@@ -25,9 +25,9 @@ test_that("Test checking for mba file", {
 })
 
 test_that("Test detecting format", {
-  plate1_filepath <- system.file("extdata", "CovidOISExPONTENT.csv", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
-  plate2_filepath <- system.file("extdata", "CovidOISExPONTENT_CO.csv", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
-  plate3_filepath <- system.file("extdata", "random_intelliflex.csv", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
+  plate1_filepath <- system.file("extdata", "CovidOISExPONTENT.csv", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
+  plate2_filepath <- system.file("extdata", "CovidOISExPONTENT_CO.csv", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
+  plate3_filepath <- system.file("extdata", "random_intelliflex.csv", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
 
   expect_equal(detect_mba_format(plate1_filepath), "xPONENT")
   expect_equal(detect_mba_format(plate2_filepath), "xPONENT")
@@ -36,7 +36,7 @@ test_that("Test detecting format", {
 })
 
 test_that("Test obtaining an output directory", {
-  plate_filepath <- system.file("extdata", "multiplate_mock", "other", "random2_xponent.csv", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
+  plate_filepath <- system.file("extdata", "multiplate_mock", "other", "random2_xponent.csv", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
   input_dir <- fs::path_dir(plate_filepath)
   input_dir_parent <- fs::path_dir(input_dir)
   specified_output_dir <- fs::path(tempdir(check = TRUE))
@@ -64,7 +64,7 @@ test_that("Test obtaining an output directory", {
 
 
 test_that("Test processing a mock directory", {
-  dir <- system.file("extdata", "multiplate_mock", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
+  dir <- system.file("extdata", "multiplate_mock", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
   expect_no_error(capture.output(
     process_dir(dir, dry_run = T, recurse = T, flatten_output_dir = T)
   ))
@@ -90,14 +90,14 @@ test_that("Test processing a mock directory", {
 })
 
 test_that("Test processing a directory with a single plate", {
-  dir <- system.file("extdata", "multiplate_lite", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
+  dir <- system.file("extdata", "multiplate_lite", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
   output_dir <- tempdir(check = TRUE)
   plates <- process_dir(dir, return_plates = T, output_dir = output_dir)
   expect_length(plates, 2)
 })
 
 test_that("Test processing a reallife directory with merge output", {
-  dir <- system.file("extdata", "multiplate_reallife_reduced", package = "SerolyzerR", mustWork = TRUE) # get the filepath of the csv dataset
+  dir <- system.file("extdata", "multiplate_reallife_reduced", package = "SerolyzeR", mustWork = TRUE) # get the filepath of the csv dataset
   output_dir <- tempdir(check = TRUE)
 
   # Clean up the tmp directory